Upgrade System Framework
Addition of the framework for the upgrade system for the train. The position of the table is only for testing and not the final position. There is only one upgrade that is done on its own at the moment. When the buttons have been implemented the script that is different from the crating table will be commented