Change to damage and train collision
Changed the enemy so if it takes damage and its hp is 0 or lower it dies, drops are not enabled though. If the train runs into an enemy they take damage depending on train speed. All items must have PickUpAble tag due to another change in physics stopper fixing the error of it trying to access items it can't do anything with.